Hide JENKINS-57326 - don't create bogus jgit
When the git section lists "home" as "jgit", it will create a command
line git implementation named "jgit".  That is not what is intended by
the demonstration and will mislead users experimenting with the
repository. Better to hide the bug for how so that users are not
confused when they see a git implementation named "jgit" which does
not use JGit.